Clean up 3 Method Ecosystem and define Proof Fetching
This commit has two aspects.
1) It cleans up the 3-method ecosystem section and eliminates many fo the outstanding questions
2) In doing so I needed to clearly define 'Proof Fetching which is the aspect of STH Pollination that has a client get inclusion proofs. Clients can pollinate STHs without doing proof fetching, now we define and explain this better.

Proof Fetching is made generic, because now that people are talking about including Inclusion proofs in certificates, a client may have a historical STH that they will need to resolve to a recent STH.  Requires a proof lookup, but now it's a consistency proof, not inclusion proof.